Issue #798: Enable warnings as errors for Core for windows.
Fix all MSVC C4244 build warnings and enable warnings as errors

This comprehensive fix addresses all C4244 narrowing conversion warnings
across the codebase and enables warnings-as-errors for all targets.

Major changes:
- Enable warnings-as-errors compiler flags (MSVC /WX, GCC/Clang -Werror)
- Fix pragma warnings: add GCC/Clang pragmas, fix MSVC pragma syntax
- Fix missing return statements in operator<=> and Animation::GetVariables
- Add [[nodiscard]] attributes where return values should not be ignored
- Remove unused variables and parameters throughout codebase
- Initialize struct members in CalChartAnimationCommand to prevent warnings
- Add explicit static_cast for narrowing conversions (float→unsigned, double→int, size_t casts)
- Add SetVarValue(double) overload to minimize casts in continuity code
- Fix SetupInstruments iterator to work with std::ranges::subrange
- Add explicit float casts in CalChartPrintShowToPS for double→float conversions

Files modified (23 files):
- cmake/compiler.cmake: Enable /WX and -Werror
- core/CalChartAnimation.cpp: Fix operator<=> return
- core/CalChartAnimationCommand.h: Initialize struct members
- core/CalChartAnimationCompile.h/cpp: Add SetVarValue(double) overload
- core/CalChartConfiguration.cpp: static_cast Radian values to float
- core/CalChartPoint.cpp: Integer division for radius calculations
- core/CalChartPrintShowToPS.cpp: Explicit float casts for coordinate conversions
- core/CalChartSheet.cpp: static_cast std::distance results
- core/CalChartAnimationSheet.cpp: static_cast std::distance results
- core/e7_transition_solver.cpp: static_cast for SolverCoord conversions
- core/viewer_translate.cpp: static_cast angle return to float
- src/animate.cpp: Remove unused parameters
- src/BackgroundImages.cpp: Add GCC/Clang pragmas for -Wimplicit-fallthrough
- src/CalChartDrawPrimitivesHelper.cpp: Add [[nodiscard]], fix pragma syntax
- src/CalChartFrame.cpp: Remove unused variables
- src/CalChartView.cpp: Add GCC/Clang pragmas
- src/ContinuityBrowserPanel.cpp: Remove unused variables
- src/FieldView.cpp: Fix GetVariables() return statement
- src/PreferencesContCellSetup.cpp: Remove unused parameter
- src/SetupInstruments.cpp: Fix iterator traits for std::ranges::subrange
- src/modes/CalChartModeDrawing.cpp: Remove unused variables
- src/modes/CalChartModeEdit_ManipulatePointsHelper.cpp: Remove unused variables

All tests pass. No warnings in macOS build (Clang). Ready for Windows CI validation.

Providing different levels to the warnings as errors.

+++ core/CalChartAnimationSheet.cpp
@@ -226,7 +226,7 @@ auto Sheets::BeatToSheetOffsetAndBeat(Beats beat) const -> std::tuple<size_t, Be
     if (where == mRunningBeatCount.end()) {
         return { mRunningBeatCount.size(), beat - TotalBeats() };
     }
-    auto index = std::distance(mRunningBeatCount.begin(), where);
+    auto index = static_cast<size_t>(std::distance(mRunningBeatCount.begin(), where));
     return { index, beat - (*where - mSheets.at(index).GetNumBeats()) };
 }
